backend/requirements-dev.txt
pytest~=7.2.1
black~=23.1.0
coverage~=7.10
pre-commit~=3.0.3
pylint~=2.16.1
pylint-django~=2.5.3
